datalayer (aka repository) sans DB

+++ product-service/src/main/java/com/example/product_service/controller/ProductController.java
@@ -22,7 +22,7 @@ public class ProductController {
   }
 
   @GetMapping("/{id}")
-  public Product getProductById(@PathVariable int id) {
+  public Product getProductById(@PathVariable Long id) {
     return productService.getProductById(id);
   }
